PHP-05 -- ex13: tests nécéssaires.

+++ ex13/src/Service/DeleteEmployeesService.php
@@ -2,17 +2,20 @@
 
 namespace App\Service;
 
-use App\Repository\EmployeeRepository;
 use Exception;
 use App\Service\UtilsTableService;
+use App\Repository\EmployeeRepository;
 use Doctrine\ORM\EntityManagerInterface;
+use App\Service\EmployeesValidationUtilsService;
 
 class DeleteEmployeesService 
 {
 	public function __construct(
         private readonly EntityManagerInterface $em,
 		private readonly EmployeeRepository $repo,
-        private readonly UtilsTableService $utilsTableService
+        private readonly UtilsTableService $utilsTableService,
+        private readonly EmployeesValidationUtilsService $validationUtils
+
 	) {}
     
     public function deleteAllTableContent(): string
@@ -49,6 +52,14 @@ class DeleteEmployeesService
 
             if (!$employee)
                 return 'danger:Employee with ID ' . $id . ' not found.';
+            if (!$this->validationUtils->canDeleteCEO($employee))
+                return 'danger:Cannot delete the CEO while other employees exist.';
+
+            if (!$this->validationUtils->canDeleteCOO($employee))
+                return 'danger:Cannot delete the COO while they manage employees.';
+
+            if (!$this->validationUtils->canDeleteManager($employee))
+                return 'danger:Cannot delete a manager who still has employees.';
             $this->em->remove($employee);
             $this->em->flush();
             return 'success:Employee deleted successfully!';
